Move azCliUtils from artifacts-common to azure-arm-rest (#340)
* Moved azCliUtils from artifacts-common to azure-arm-rest.
