+++ common-npm-packages/artifacts-common/webapi.d.ts
@@ -2,4 +2,3 @@ import * as api from 'azure-devops-node-api';
 import { IRequestOptions } from 'azure-devops-node-api/interfaces/common/VsoBaseInterfaces';
 export declare function getWebApiWithProxy(serviceUri: string, accessToken: string, options?: IRequestOptions): api.WebApi;
 export declare function getSystemAccessToken(): string;
-export declare function getFederatedToken(connectedServiceName: string): Promise<string>;
