+++ common-npm-packages/azure-arm-rest/package.json
@@ -1,6 +1,6 @@
 {
   "name": "azure-pipelines-tasks-azure-arm-rest",
-  "version": "3.242.0",
+  "version": "3.242.1",
   "description": "Common Lib for Azure ARM REST apis",
   "repository": {
     "type": "git",
